# documentation/swagger.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8080'
basePath: /
...
paths:
  /api/v1/parts:
    post:
      tags:
        - part-controller
      summary: Create a new part
      operationId: Create Part
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: part
          description: Information of the part to be crated
          required: true
          schema:
            $ref: '#/definitions/PartDTO'
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '402':
          description: |
            BAD REQUEST: 
            * the part code xxxxxxxx already exist. 
            * not found a provider with this id. 
            * not found a discount rate with this id. 
        '403':
          description: FORBIDDEN
        '404':
          description: Not Found
      deprecated: false
    put:
      tags:
        - part-controller
      summary: Update a part
      operationId: Update Part
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: part
          description: |-
            Information of the part to be update 
            * partCode is required. 
            * only the attributes sent as parameters will be modified. 
            * makerId should exist. 
            * discountId should exist.
          required: true
          schema:
            $ref: '#/definitions/PartDTO'
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '402':
          description: |-
            BAD REQUEST: 
            * no parameters were received to update. 
            * not found a provider with this id 
            * not found a discount rate with this id
        '403':
          description: FORBIDDEN
        '404':
          description: |-
            NOT FOUND: 
            * part with code xxxxxxxx not exist
      deprecated: false
...
  /api/v1/parts/discountRates:
    get:
      tags:
        - basic-aux-crud-controller
      summary: getALLDiscountRate
      operationId: getALLDiscountRateUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/DiscountRateDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    post:
      tags:
        - basic-aux-crud-controller
      summary: addDiscountRate
      operationId: addDiscountRateUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: discountRateDTO
          description: discountRateDTO
          required: true
          schema:
            $ref: '#/definitions/DiscountRateDTO'
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/api/v1/parts/discountRates/{id}':
    get:
      tags:
        - basic-aux-crud-controller
      summary: findDiscountById
      operationId: findDiscountByIdUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/DiscountRateDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
...
  /api/v1/parts/providers:
    get:
      tags:
        - basic-aux-crud-controller
      summary: findAllProviders
      operationId: findAllProvidersUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/ProviderDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    post:
      tags:
        - basic-aux-crud-controller
      summary: addProvider
      operationId: addProviderUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: providerDTO
          description: providerDTO
          required: true
          schema:
            $ref: '#/definitions/ProviderDTO'
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/api/v1/parts/providers/{id}':
    get:
      tags:
        - basic-aux-crud-controller
      summary: findProviderById
      operationId: findProviderByIdUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ProviderDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
...
definitions:
...
  DiscountRateDTO:
    type: object
    properties:
      idDiscountRate:
        type: integer
        format: int64
        example: 54321
        description: Unique identifier of the DiscountRate. Each discount has its own id
      description:
        type: string
        example: Discount for part Espolon BMW 320i
        description: Description of the DiscountRate
      discount:
        type: string
        example: '%30'
        description: Value of the Discount
    title: DiscountRateDTO
...
  PartDTO:
    type: object
    required:
      - description
      - discountId
      - longDimension
      - makerId
      - netWeight
      - normalPrice
      - partCode
      - quantity
      - salePrice
      - tallDimension
      - urgentPrice
      - widthDimension
    properties:
      partCode:
        type: integer
        format: int32
        example: 98521789
        description: Unique identifier of the Part.Two Parts cant have the same partId.
      quantity:
        type: integer
        format: int32
        example: 1
        description: Quantity of the part
        minimum: 1
        exclusiveMinimum: false
      description:
        type: string
        example: Espolon BMW 320i
        description: Description of the part.
      netWeight:
        type: integer
        format: int32
        example: 64
        description: NetWeight of the part
        minimum: 1
        exclusiveMinimum: false
      longDimension:
        type: integer
        format: int32
        example: 69
        description: Long dimension of the part
        minimum: 1
        exclusiveMinimum: false
      widthDimension:
        type: integer
        format: int32
        example: 20
        description: Width dimension of the part
        minimum: 1
        exclusiveMinimum: false
      normalPrice:
        type: number
        format: double
        example: 20
        description: 'Normal price of the part, min 1'
        minimum: 1
        exclusiveMinimum: false
      tallDimension:
        type: integer
        format: int32
        example: 30
        description: Tall dimension of the part
        minimum: 1
        exclusiveMinimum: false
      salePrice:
        type: number
        format: double
        example: 20
        description: 'Sale price of the part, min 1'
        minimum: 1
        exclusiveMinimum: false
      urgentPrice:
        type: number
        format: double
        example: 90
        description: 'Urgent price of the part, min 1'
        minimum: 1
        exclusiveMinimum: false
      makerId:
        type: integer
        format: int64
        example: 1
        description: 'Id of maker, should exist'
      discountId:
        type: integer
        format: int64
        example: 1
        description: 'Id of discount, should exist'
    title: PartDTO
    description: |-
      Basic fields to create or update a Part 
      with validation
...
  ProviderDTO:
    type: object
    required:
      - address
      - country
      - name
      - phone
    properties:
      idProvider:
        type: integer
        format: int64
        example: 93843
        description: id of the part provider
      name:
        type: string
        example: BMW
        description: Name of the part provider
      address:
        type: string
        example: 'Street 123, Office 205'
        description: Address of the part provider
      phone:
        type: string
        example: +1 567 8955
        description: Phone number of the part provider
      country:
        type: string
        example: Nigeria
        description: Country of the part provider
    title: ProviderDTO
    description: Basic fields of a provider
...